+++ src/Server.php
@@ -3,8 +3,9 @@
 namespace Rubix\Server;
 
 use Rubix\ML\Estimator;
+use Psr\Log\LoggerAwareInterface;
 
-interface Server
+interface Server extends LoggerAwareInterface
 {
     /**
      * Serve a model.
@@ -12,4 +13,18 @@ interface Server
      * @param \Rubix\ML\Estimator $estimator
      */
     public function serve(Estimator $estimator) : void;
+
+    /**
+     * Return the number of requests that have been received.
+     *
+     * @var int
+     */
+    public function requests() : int;
+
+    /**
+     * Return the uptime of the server in seconds.
+     *
+     * @return int
+     */
+    public function uptime() : int;
 }