+++ src/Traits/LoggerAware.php
@@ -32,14 +32,4 @@ trait LoggerAware
     {
         $this->logger = $logger;
     }
-
-    /**
-     * Return if the logger is logging or not.
-     *
-     * @var bool
-     */
-    public function logging() : bool
-    {
-        return isset($this->logger);
-    }
 }