+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - RPC client now has exponential backoff mechanism
     - Fixed JSON serialization of Estimator and Data types
     - Rename Rank and RankSample to Score and ScoreSample
+    - Move Verbose interface to Server
 
 - 0.0.2-beta
     - Changed name of Binary serializer to Igbinary